refactor(network): Phase0 统一注册表入口与实体整数ID+ItemId 扩 u32
网络层重写前置:为后续 VarInt 编解码铺路。

- 新增 src/common/registry/RegistryAccess:聚合引用
  Item/Block/Entity/Biome 四大注册表单例,为网络层 codec
  提供按整数 ID/资源位置查表统一入口(对应 MC Java 1.21.11
  RegistryFriendlyByteBuf 持有 RegistryAccess 的角色)。
- EntityRegistry:registerType 按注册顺序分配递增 u32 id,
  bindIdentity(name,id);新增 getTypeById(u32),解锁
  SpawnEntity VarInt 类型 ID 对齐。
- ItemId 由 u16 扩为 u32:网络层按 VarInt 编码,u32 容量足够
  且与注册表下标类型一致;ItemStack.cpp 网络写点显式收窄 u16
  并标 TODO(Phase1-3 改为 VarInt)。
- 清理 8 处冗余 static_cast<ItemId> 与 BeaconEntity 死代码
  (itemId>u16_max 判定已恒假)。
- CLAUDE.md 补基岩版源码参考路径。

+++ src/common/core/Types.hpp
@@ -96,7 +96,8 @@ using EntityInstanceId = u64;
 inline constexpr EntityInstanceId INVALID_ENTITY_ID = 0;
 
 // 物品ID类型
-using ItemId = u16;
+// 网络层按 VarInt 编码，u32 容量足够且与注册表分配的下标类型一致。
+using ItemId = u32;
 
 // 生物群系ID类型
 using BiomeId = u16;
